deep-research-py/requirements.txt
python-dotenv>=1.0.0
streamlit>=1.36.0
openai>=1.30.1
langchain>=0.2.0
langchain-openai>=0.2.0
httpx>=0.27.0
beautifulsoup4>=4.12.3
markdown2>=2.4.12
pydantic>=2.7.1
python-frontmatter>=1.0.0
redis>=5.0.1
pytest>=8.0.0
pytest-asyncio>=0.23.5
pytest-cov>=4.1.0
langchain-core>=0.2.0
firecrawl-py>=0.1.0


